[COLRv1] Visited Set needs to be carried down PaintColrGlyph
PaintColrGlyph recursively started drawing a new glyph but did not carry
over the visisted set so that a recursion was left
undetected. Solve this by moving the VisitedSet up one level and
carry it down to starting the new glyph recursion when encountering
a PaintColrGlyph table.

Bug: chromium:1274206
Change-Id: Ic1dcc90ff285d0e802125aff02b9d2097d8e5952
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/477096
Commit-Queue: Ben Wagner <bungeman@google.com>
Reviewed-by: Ben Wagner <bungeman@google.com>