Correct documentation on setting glyph.fMaskFormat
It used to be that SkScalerContext::generateMetrics could not reliably
set the fMaskFormat and generateImage could modify it. This has since
changed and the code not generally depends on generateMetrics properly
setting fMaskFormat and generateImage not changing it.

Change-Id: I9723d71370de587ef3968624d4a7eebed134b321
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/477816
Commit-Queue: Ben Wagner <bungeman@google.com>
Auto-Submit: Ben Wagner <bungeman@google.com>
Reviewed-by: Herb Derby <herb@google.com>