Geometry Processor can output local coord in fragment stage.
Currently GPs inform the program builder of the variable that contains
the local coordinates and it is assumed to be defined in the VS.
Additionally, the builder process will attempt to lift derived
coordinate computations to the VS as varying outputs when they are a
linear/projective transformation of the the GP's local coords.

This change makes it so that the GP can state that the local coord
variable is instead available in the fragment stage. If the GP uses
this option then lifting of coordinate transformations to the VS is
blocked.

Bug: skia:12720
Change-Id: I63d3ffe069c9e4ac43ac5bb305d6e04a11119626
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/481279
Reviewed-by: Michael Ludwig <michaelludwig@google.com>
Commit-Queue: Brian Salomon <bsalomon@google.com>