Disable SkVMBlitter cache for Programs with trace hooks.
The trace hooks don't participate in the cache key, so the "matching"
Program in the cache may not have the same trace hook pointers as the
Builders. In fact, the trace hook pointers in the cached Program are
very likely to point to objects that no longer exist at all.

We could incorporate the trace hooks into the key, but this seems like a
losing battle--they're raw pointers and are extraordinarily unlikely to
stay consistent over time. Since trace hooks are only used in a
debugging context, where performance is not very important, the simplest
solution is to just exclude programs with trace hooks from the cache
entirely.

Change-Id: I65799331da012656c8b13e76107bc893418a874c
Bug: skia:12708
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/481440
Commit-Queue: John Stiles <johnstiles@google.com>
Auto-Submit: John Stiles <johnstiles@google.com>
Reviewed-by: Brian Osman <brianosman@google.com>
Commit-Queue: Brian Osman <brianosman@google.com>