[infra] Move taskdriver for PushBazel task into skia repo
AFAIK, there was no need for it to be in the infra repo
and that made it difficult to iterate on.

It successfully checks out the skia-infra repo at ToT
and runs make bazel_release_ci in //jsfiddle.
This uploads the docker image, so a followup will
pipe it into the Pub/Sub flow.

Change-Id: I8c8fd9755003162b4ff38a7ffea97d1ea64cd493
Bug: skia:10614
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/481236
Reviewed-by: Eric Boren <borenet@google.com>