GrMemoryPool always standard_layout for offsetof
GrMemoryPool in debug has extra fields and so wants to make use of
offsetof. The offsetof can only be taken on standard layout objects.
However, in debug std::unique_ptr may not be standard layout, which
makes SkAutoTArray not standard layout, which makes SkTHashTable not
standard layout, which makes SkTHashSet not standard layout, which makes
GrMemoryPool not standard layout, producing a compile error.

Since this is debug only, move the debug only fields behind a pointer
(which unfortunately cannot be a std::unique_ptr, since it may not be
standard layout). This allows building with recent libc++ headers and
clang.

Change-Id: Id9d312d7939808399d0796428de218d30263c26f
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/482004
Reviewed-by: Michael Ludwig <michaelludwig@google.com>
Commit-Queue: Ben Wagner <bungeman@google.com>