Clean up intrinsics slightly
We want to distinguish between actual intrinsics (functions declared in
the SkSL includes, but defined by the generators), and builtin functions
(declared and defined in the includes).

This removes unpremul from the intrinsic list (it's a builtin function,
not an intrinsic), and adds an assert that no intrinsic is defined.

Change-Id: I7b4f0cefd98364fe71aa570357f0baed3f5887ff
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/484441
Reviewed-by: John Stiles <johnstiles@google.com>
Commit-Queue: Brian Osman <brianosman@google.com>
Auto-Submit: Brian Osman <brianosman@google.com>