better handling of atlas size constraints
The problem is using the dimensions of the path glyphs to
approximate what will happen when we scale things to fit in
the atlas constraints. The path dimensions do necessarily scale to
represent the RGB dimensions of the emoji.

Round the max dimension up to provide a better margin of error.

The better solution will be to calculate the maxDimension in
MakeMask instead of MakePath, but this is a bigger engineering
exercise.

Bug: chromium:1280180

Change-Id: I994ed2b3461677890418d0472f31bd1b2651b037
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/485216
Reviewed-by: Ben Wagner <bungeman@google.com>
Commit-Queue: Herb Derby <herb@google.com>