Use clipped bounds for tessellation heuristic
The heuristic estimates the fill cost of the path based on its bounds.
The goal is to use the inner triangulator for large paths with a low
vertex count so that filling is simpler (no stencil), but triangulation
cost is manageable.

Considering the unclipped bounds means that very large path coordinates
can mislead the heuristic into thinking there is substantial fill cost,
when in fact that's always bounded by the clip dimensions.

Bug: skia:12764
Change-Id: I9d14bbdd2b35df121b6a55d4a278656bf16ae8eb
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/488528
Reviewed-by: Greg Daniel <egdaniel@google.com>
Commit-Queue: Michael Ludwig <michaelludwig@google.com>