ANDROIDFRAMEWORK_DEBUGF -> RENDERENGINE_ABORTF
Bug: b/206415266

We added ANDROIDFRAMEWORK_DEBUGF logs to help diagnose a crash, but the
bugreports start with the crash, which occurs after these logs. Change
the logs to ABORTs so the bugreports will contain this new information.

Only turn on the new abort statements when in RenderEngine. Add a new
define so that we can distinguish between the version of Skia used in
RenderEngine from the one in HWUI.

Change-Id: Ib096fc0b09741ec73ddedaa7cfdd21c5584d74cd
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/489476
Reviewed-by: Derek Sollenberger <djsollen@google.com>
Commit-Queue: Leon Scroggins <scroggo@google.com>