Convert RENDERENGINE_ABORTF to LOG_ALWAYS_FATAL
Bug: b/213293147
Bug: b/206415266

The intent of RENDERENGINE_ABORTF was for the error message to show as
part of the stack trace, as logs before the stack trace may not be
available. However, a closer reading of SK_ABORTF reveals that it's just
an SkDebugf followed by an abort, and a device hitting the abort still
doesn't show the debug statement.

Switch to LOG_ALWAYS_FATAL, which includes the message in the stack
trace.

Change-Id: I7b83805d946011f39fd43ac8efe742acf760c769
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/491977
Reviewed-by: Derek Sollenberger <djsollen@google.com>
Commit-Queue: Leon Scroggins <scroggo@google.com>