[infra] Add a Debug-All-NativeFonts CPU job for all OS versions on the new Mac Mini 7.1s.
These Minis have a new CPU (I5-4278U) we've never cared about before, so
add that too. (We had some other Minis of the same model in the Golo,
but we didn't care about their CPUs because we were running only GPU
jobs on them.)

If you want to see a good diff of those, use FileMerge or something that
uses a similar algorithm; git makes a total hash of it.

Change-Id: Id7c4f6bc06f83d2ed4a0034e9174b9e69d3d15b2
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/495424
Reviewed-by: Ben Wagner <bungeman@google.com>
Commit-Queue: Erik Rose <erikrose@google.com>