Rewrote command line flag parsing in SkSLMain
The previous command line parsing was extremely limited and did not
allow us to easily add additional flags. This generalizes the parse so
that we are not dependent on fixed numbers of arguments in fixed
positions.

In the future we may be able to use CommandLineFlags, but right now it
doesn't completely support our needs.

Change-Id: I55bdc993e9ea71a85678d11b991dfc69e151b014
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/496098
Reviewed-by: John Stiles <johnstiles@google.com>
Commit-Queue: Ethan Nicholas <ethannicholas@google.com>