Fixed an issue with SkSL For loop symbol tables
When transforming a for loop with multiple variable initializers, e.g.
for (int x = 0, y = 0; ...), we rewrite it to:
{ int x = 0; int y = 0; for (;...)

But we would keep the symbol table containing x and y attached to the
for loop. This means that while walking the code tree, we would run
into the declarations for x and y before running into the symbol table
containing them. This poses problems for the dehydrator, which would
fail when running into this construct. This fix simply moves the
symbol table to the containing block instead.

Change-Id: Ic806bbda19c954224a726b4a315f3102852e26b1
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/494717
Reviewed-by: John Stiles <johnstiles@google.com>
Commit-Queue: Ethan Nicholas <ethannicholas@google.com>